# swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "version": "1.0.0",
        "title": "REST API",
        "description": ""
    },
    "host": "localhost:3000",
    "basePath": "/",
    "schemes": [
        "http"
    ],
...
    "paths": {
...
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cria um login para usuário.",
                "description": "Cria um novo login que será usado para acesso ao sistema.",
                "requestBody": {
                    "description": "Dados para criação do login",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterCommonUsers"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "token"
                                    ],
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhb.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIi.lpBikrwO7XOIIAfH"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou incompletos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/admin": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cria um login para um administrador.",
                "description": "Cria um novo login de administrador que será usado para acesso ao sistema restrito.",
                "requestBody": {
                    "description": "Dados para criação do login",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterAdminUsers"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "token"
                                    ],
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhb.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIi.lpBikrwO7XOIIAfH"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos ou incompletos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/authenticate": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Autentica um usuário que ainda não é cliente.",
                "description": "Autentica um usuário, caso o usuário não existe, ou a combinação de usuário e senha, retorna erro.",
                "requestBody": {
                    "description": "Dados para autenticação do usuário",
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "type": "object",
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "tigas@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "qwerty123"
                                    }
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "required": [
                                        "token"
                                    ],
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhb.eyJ0b2tlbl90eXBlIjoiYWNjZXNzIi.lpBikrwO7XOIIAfH"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Autenticação inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Erro"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "RegisterCommonUsers": {
                "required": [
                    "firstName",
                    "lastName",
                    "email",
                    "password",
                    "birthdate"
                ],
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string",
                        "description": "Campo texto de preenchimento obrigatório. Inserir primeiro nome do usuário.",
                        "maxLength": 200,
                        "example": "Joaquim"
                    },
                    "lastName": {
                        "type": "string",
                        "description": "Campo texto de preenchimento obrigatório. Inserir último nome do usuário.",
                        "maxLength": 200,
                        "example": "Xavier"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "teste@gmail.com"
                    },
                    "birthdate": {
                        "type": "string",
                        "example": "21/12/2000"
                    },
                    "password": {
                        "type": "string",
                        "example": "210491Ag"
                    }
                }
            },
            "RegisterAdminUsers": {
                "required": [
                    "firstName",
                    "lastName",
                    "email",
                    "password",
                    "birthdate"
                ],
                "type": "object",
                "properties": {
                    "firstName": {
                        "type": "string",
                        "description": "Campo texto de preenchimento obrigatório. Inserir primeiro nome do usuário.",
                        "maxLength": 200,
                        "example": "Joaquim"
                    },
                    "lastName": {
                        "type": "string",
                        "description": "Campo texto de preenchimento obrigatório. Inserir último nome do usuário.",
                        "maxLength": 200,
                        "example": "Xavier"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "teste@gmail.com"
                    },
                    "birthdate": {
                        "type": "string",
                        "example": "21/12/2000"
                    },
                    "password": {
                        "type": "string",
                        "example": "210491Ag+"
                    }
                }
            },
...
            "Erro": {
                "description": "Mensagem de erro detalhada",
                "properties": {
                    "mensagem": {
                        "type": "object",
                        "description": "Conteúdo da mensagem de erro."
                    }
                }
            }
        }
    }
}